requirements.txt
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.0.3
numpy==1.17.4
opencv-python==4.1.2.30
Pillow==6.2.0
pkg-resources==0.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
six==1.14.0
torch==1.0.0
torchvision==0.2.1
dlib==19.19.0